% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{featureSelection}
\alias{featureSelection}
\alias{print.feature_select}
\alias{update.feature_select}
\title{Feature Selection Object Declaration}
\usage{
featureSelection(
  data,
  candidate.markers,
  model.type,
  search.type,
  runs = 1,
  folds = 1,
  cost = c("AUC", "R2", "CCC", "MSE", "sens", "spec"),
  keep.models = FALSE,
  bootstrap = FALSE,
  stratified = FALSE,
  strat.column = NULL,
  random.seed = sample(1000, 1)
)

\method{print}{feature_select}(x, ...)

\method{update}{feature_select}(object, ...)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} class object containing
features and clinical data to be used for modeling.}

\item{candidate.markers}{Character. List of candidate markers, i.e. columns
names, from the data object.}

\item{model.type}{An instantiated \code{modelType} object, generated via a call
to one of the \code{\link[=modelType]{modelType()}} functions.}

\item{search.type}{An instantiated \code{searchType} object, generated via a
call to one of the \code{\link[=searchType]{searchType()}} functions.}

\item{runs}{Integer. How many runs to perform.}

\item{folds}{Integer. How many fold cross-validation to perform.}

\item{cost}{Character. A string to be used in defining the cost
function. Currently one of:
\describe{
\item{\code{AUC}}{Area Under the Curve}
\item{\code{MSE}}{Mean-Squared Error}
\item{\code{CCC}}{Concordance Correlation Coefficient}
\item{\code{R2}}{R-squared - regression models}
\item{\code{sens} or \code{spec}}{Sensitivity + Specificity}
}}

\item{keep.models}{Logical. Should model objects be retained in the
cross validation sub-list? This eats up memory (\code{default = FALSE}).}

\item{bootstrap}{Logical. Should data be bootstrapped rather
than set up in cross-validation folds? The result is multiple runs (defined
by runs) with 1 Fold each. The full dataset will be sampled with replacement
to generate a training set of equivalent size. The samples not chosen during
sampling make up the test set.}

\item{stratified}{Logical. Should cross-validation or bootstrap
folds be stratified based upon the stratification column specified in
\code{strat.column}?}

\item{strat.column}{Character. Which column to use for stratification of
cross-validation or bootstrap folds. If \code{NULL} (default), column name
\code{"Response"} will be used and thus must be present.}

\item{random.seed}{Integer. Used to control the random number generator for
reproducibility.}

\item{x}{An object of class \code{feature_select} (list) from a call
call to \code{\link[=featureSelection]{featureSelection()}}.}

\item{...}{Arguments declared for update in \code{argument = value} format.
Arguments from the original call are preserved. Alternatively, a pass through
for the deprecation of FSMCP -> featureSelection (will eventually be removed).}

\item{object}{An object of class \code{feature_select}.}
}
\value{
A \code{"feature_select"} class object and list containing:
\item{data}{The original feature data to use.}
\item{candidate.markers}{The list of candidate features.}
\item{model.type}{A list containing model type variables of the
appropriate class for the desired model type.}
\item{search.type}{A list containing search type variables of the
appropriate class for the desired search type.}
\item{cost}{A string of the type of cost function.}
\item{cost.fxn}{A list containing cost variables of the
appropriate class for the desired object cost function.}
\item{runs}{The number of runs.}
\item{folds}{The number of folds.}
\item{keep.models}{If intermediate cross-validation models are kept?}
\item{bootstrap}{Is bootstrapping performed?}
\item{stratified}{Is cross-validation stratification performed?}
\item{strat.column}{Which field string is used in stratification.}
\item{random.seed}{The random seed used}
\item{cross.val}{A list containing the training and test indices of the
various cross validation folds.}
\item{search.complete}{Logical if the object has completed a search}
\item{call}{The original matched call.}
}
\description{
Declares and generates a \code{feature_selection} class object within the
Feature Selection framework. This object acts as the holder of data
(bootstrapped or cross-validation folds), model type, search type,
cost function, and an underlying data structure for use by other functions.

The S3 print method for objects of class \code{feature_select}.

The S3 update method allows for modification of existing
\code{feature_select} objects on-the-fly.
}
\examples{
# Simulated Test Data
data <- splyr::sim_adat

# Setup response variable
data$class_response <- factor(data$class_response)

mt <- modelType_glm(response = "class_response")
cf <- "sens"
sm <- searchType_forwardModel(15, display.name = "FeatureSelection Plot")
ft <- featureselectr:::getAnalytes(data) # select candidate features
fs <- featureSelection(data, candidate.markers = ft,
                       model.type = mt, search.type = sm, cost = cf,
                       strat.column = "class_response", runs = 5, folds = 5)
# S3 Print method
fs

# Using the S3 Update method to modify existing FS object:
# change model type, cost function, and random seed
fs2 <- update(fs, model.type = modelType_nb(),
              cost = "AUC", random.seed = 99)
fs2

# change number of runs & folds
# requires re-calculation of cross-validation parameters
fs3 <- update(fs, runs = 20, folds = 10)
fs3

}
\references{
Hastie, Tibshirani, and Friedman. Elements of Statistical Learning:
Data Mining, Inference , and Prediction. 2nd Edition. Springer. 2009.
}
\seealso{
\code{\link[=Search]{Search()}}
}
\author{
Kirk DeLisle, Stu Field
}
