% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureselectr-package.R
\docType{package}
\name{featureselectr-package}
\alias{featureselectr}
\alias{featureselectr-package}
\title{featureselectr: Perform Feature Selection/Optimization}
\description{
The feature selection package contains functionality designed for feature selection, model building, and/or classifier development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stufield.github.io/featureselectr}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kirk DeLisle \email{rkdelisle@gmail.com}
}

}
\keyword{internal}
