% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFSmarkers.R
\name{getFSmarkers}
\alias{getFSmarkers}
\title{Get Feature Selection Markers}
\usage{
getFSmarkers(x)
}
\arguments{
\item{x}{A \code{feature_select} search object.}
}
\value{
A list containing:
\item{max.markers}{Combination of SOMAmers that gives maximum/minimum
mean cost function.}
\item{markers.1se.from.max}{Combination of SOMAmers that has a mean
cost function that is one standard error (SE) from the
maximum/minimum mean cost function.}
\item{markers.2se.from.max}{Combination of SOMAmers that has a mean
cost function that is 1.96*SE from the maximum/minimum mean
cost function.}
}
\description{
Function to return the maximum, se1, and se2 markers from either
forward or backward selection from a \code{feature_select} class object.
}
\examples{
data  <- splyr::sim_adat
feats <- attributes(data)$sig_feats$class
fs <- featureSelection(data, candidate.markers = feats,
                       search.type = searchType_forwardModel(),
                       model.type = modelType_glm("class_response"),
                       stratified = TRUE,
                       cost = "AUC", runs = 2, folds = 2)
fs_obj <- Search(fs)
getFSmarkers(fs_obj)
}
\author{
Stu Field
}
