% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-type.R
\name{model_type}
\alias{model_type}
\alias{model_type_lr}
\alias{model_type_nb}
\alias{model_type_lm}
\title{Create Model Type Function Definitions}
\usage{
model_type_lr(response = "Response")

model_type_nb(response = "Response")

model_type_lm(response = "Response")
}
\arguments{
\item{response}{\code{character(1)}. The string of the column name to use as the
response variable. Assumed to be a binary factor type.}
}
\value{
An object of the appropriate class according to the
model type chosen, one of: \code{fs_lr}, \code{fs_nb}, or \code{fs_lm}.
}
\description{
Declares a model type the Feature Selection framework. There are
currently 3 model type options:
\itemize{
\item Logistic Regression Models (lr)
\item Linear Regression Models (lm)
\item Naive Bayes Models (nb)
}
Typically these functions can be called with their defaults,
especially if called upon a \code{tr_data} class object (see \pkg{libml}).
}
\examples{
# Logistic Regression
model_type_lr()

# Robust Parameter Naive Bayes (default)
model_type_nb()

# Linear Regression
model_type_lm()

}
\author{
Kirk DeLisle, Stu Field
}
