% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTypes.R
\name{modelType}
\alias{modelType}
\alias{modelType_glm}
\alias{modelType_nb}
\alias{modelType_lm}
\title{Create Model Type Function Definitions}
\usage{
modelType_glm(response = "Response")

modelType_nb(response = "Response")

modelType_lm(response = "Response")
}
\arguments{
\item{response}{Character. The string of the column name to use as the
response variable. Assumed to be two-class, factor type.}
}
\value{
An object of the appropriate class according to the
model type chosen, one of: \code{fs_glm}, \code{fs_nb}, or \code{fs_lm}.
In addition, key-value pairs in a list containing:
\item{response}{See function arguments above.}
}
\description{
Declares a model type the Feature Selection framework. There are
currently 3 model type options:
\itemize{
\item Logistic Regression Models (glm)
\item Linear Regression Models (lm)
\item Naive Bayes Models (nb)
}
Typically these functions can be called with their defaults, especially if
called upon a data set of class \code{tr_data} (See Classify).

Naive Bayes Models

Linear Regression Models
}
\examples{
# Logistic Regression
modelType_glm()

# Robust Parameter Naive Bayes (default)
modelType_nb()

# Linear Regression
modelType_lm()

}
\author{
Kirk DeLisle, Stu Field
}
