% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-type-backward-model.R,
%   R/search-type-forward-model.R
\name{search_type_backward_model}
\alias{search_type_backward_model}
\alias{search_type}
\alias{search_type_forward_model}
\title{Create Model Search Definitions for S3 Methods}
\usage{
search_type_backward_model(display_name = "Backward Stepwise Model Search")

search_type_forward_model(
  display_name = "Forward Stepwise Model Search",
  max_steps = 20L
)
}
\arguments{
\item{display_name}{\code{character(1)}. A title used by S3 plot methods.}

\item{max_steps}{\code{integer(1)}. Maximum number of features allowed in the model.}
}
\value{
A list containing:
\item{display_name}{The official "Title" to be used by downstream plot
methods called on the \code{feature_select} object.}
\item{max_steps}{Maximum model steps to search. Forward search only!}
}
\description{
Functions for creating model type definitions for objects in the Feature
Selection framework. Current available search types/methods are:
\itemize{
\item forward stepwise model
\item backward stepwise model
}
This information is called within the object creation to modify the its
attributes (class) to reflect the appropriate search values.
}
\section{Functions}{
\itemize{
\item \code{search_type_backward_model()}: Backward model selection search.

\item \code{search_type_forward_model()}: Forward model selection search.

}}
\examples{
search_type_forward_model()                    # the default = 20L

search_type_forward_model(max_steps = 15L)     # set to 15

search_type_forward_model("My Forward Search") # set title
}
\author{
Stu Field, Kirk DeLisle
}
