% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchType_backwardModel.R,
%   R/searchType_forwardModel.R
\name{searchType_backwardModel}
\alias{searchType_backwardModel}
\alias{searchType}
\alias{searchType_forwardModel}
\title{Create Model Search Definitions for S3 Methods}
\usage{
searchType_backwardModel(display.name = "Backward Stepwise Model Search")

searchType_forwardModel(
  max.steps = 20,
  display.name = "Forward Stepwise Model Search"
)
}
\arguments{
\item{display.name}{Character. A title or display name to use by S3 plot
generics.}

\item{max.steps}{Maximum number of covariates allowed into the model.}
}
\value{
A list containing:
\item{max.steps}{Maximum model steps to search. If forward search implemented.}
\item{display.name}{The official "Display Title" to be used by any plot
methods called on the object.}
}
\description{
Functions for creating model type definitions for objects in the Feature
Selection framework. Current available search types/methods are:
\itemize{
\item forward stepwise model
\item backward stepwise model
\item forward stepwise parameter (deprecated)
}
This information is called within the object creation to modify the its
attributes (class) to reflect the appropriate search values.
}
\section{Functions}{
\itemize{
\item \code{searchType_backwardModel()}: Backwards model selection search

\item \code{searchType_forwardModel()}: Forwards model selection search

}}
\examples{
searchType_forwardModel()                 # the default=20
searchType_forwardModel(max.steps = 15)   # set to 15
searchType_forwardModel(15, "My Awesome Forward Search")   # set title
}
\author{
Kirk DeLisle & Stu Field
}
