% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-cross.R
\name{setup_cross}
\alias{setup_cross}
\alias{setup_cross_strat}
\alias{plot_cross}
\title{Set up Cross-Validation Folds}
\usage{
setup_cross(x)

setup_cross_strat(x)

plot_cross(x)
}
\arguments{
\item{x}{A \code{feature_select} class object from a call
call to \code{\link[=feature_selection]{feature_selection()}}.}
}
\description{
These S3 methods set up the internal cross-validation.
Methods exist for both \emph{stratified} and \emph{non-stratified} scenarios.
Since these are mostly internal methods, documentation is minimal.

The \emph{stratified} version is a S3 method that sets up the
internal cross-validation for stratified folds.
Stratified implies that the structure (proportions) of the response
variable is maintained (i.e. the proportion of disease samples in the
fold is comparable to the disease proportion in the full training set).

The plotting routine allows one to visually check
the cross-validation stratification for bias.
}
\author{
Stu Field
}
