% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost}
\alias{cost}
\title{Feature Selection Cost Functions}
\usage{
cost(x)
}
\arguments{
\item{x}{A \code{feature_select} class object.}
}
\value{
A metric, a numeric value, to be optimized depending
on the type, one of AUC, CCC, MSE, R2, Sens + Spec.
}
\description{
The Feature Selection framework provides for generic cost functions to be
defined. This S3 method allows for model fit/cost to be calculated for
appropriately structured objects and requires an input object that is
structured appropriately.
This functionality is performed internally within the feature selection
algorithm and is unlikely to require direct user calls.
}
\author{
Kirk DeLisle, Stu Field
}
