% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{cost}
\alias{cost}
\title{Feature Selection Cost Functions}
\usage{
cost(x)
}
\arguments{
\item{x}{An object of class \code{feature_select} (\code{typeof = list}) from a call
call to \code{\link[=feature_selection]{feature_selection()}}.}
}
\value{
The return value is one of: Area Under the Curve (AUC),
Correspondence Correlation Coefficient (CCC), Mean Squared Error (MSE),
R-squared, or Sensitivity + Specificity, for the supplied predictions
relative to the actual values.
}
\description{
The Feature Selection framework provides for generic cost functions to be
defined. This S3 method allows for model fit/cost to be calculated for
appropriately structured objects and requires an input object that is
structured appropriately (see above).
This functionality is performed internally within the feature selection
algorithm and is unlikely to require direct calls from the user.
}
\author{
Kirk DeLisle, Stu Field
}
