% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmodel.R
\name{fitmodel}
\alias{fitmodel}
\title{Model Fitting Methods of Various Forms}
\usage{
fitmodel(x, ...)
}
\arguments{
\item{x}{A \code{feature_select} class object.}

\item{...}{Additional arguments passed to the \code{fitmodel} generic S3
method, which performs the appropriate search algorithm depending on the
object class. This is typically performed internally via the \verb{frmla =}
argument which is used to create a formula prior to being passed to one of:
\code{\link[=fit_logistic]{fit_logistic()}}, \code{\link[=fit_nb]{fit_nb()}}, \code{\link[stats:lm]{stats::lm()}}.}
}
\description{
Provides linear regression, Naive Bayes, and logistic regression
functionality within the Feature Selection framework. The model type object
requires an input object that is structured appropriately.
This functionality is performed internally within the feature selection
algorithm and is unlikely to require direct user calls.
}
\details{
The Feature Selection framework provides for generic model types to be
defined. This S3 method allows for model types to be fit for appropriately
structured objects. The model input object \emph{must} be structured
appropriately. Note the inclusion of a \code{frmla} argument
with the desired formula. Without this and in the correct form, the method
will fail. Usually it is fine to leave the defaults in place.
}
\seealso{
\code{\link[=feature_selection]{feature_selection()}}, \code{\link[=fit_logistic]{fit_logistic()}}, \code{\link[=lm]{lm()}}, \code{\link[=fit_nb]{fit_nb()}},

\code{\link[=model_type_lr]{model_type_lr()}}, \code{\link[=model_type_lm]{model_type_lm()}}, \code{\link[=model_type_nb]{model_type_nb()}}
}
\author{
Stu Field, Kirk DeLisle
}
