% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fs-features.R
\name{get_fs_features}
\alias{get_fs_features}
\title{Get Feature Selection Markers}
\usage{
get_fs_features(x)
}
\arguments{
\item{x}{A \code{feature_select} class object from a call
call to \code{\link[=feature_selection]{feature_selection()}}.}
}
\value{
A list containing:
\item{max_features}{Combination of features that gives maximum/minimum
mean cost function.}
\item{features_1se_from_max}{Combination of features that has a mean
cost function that is one standard error (SE) from the
maximum/minimum mean cost function.}
\item{features_2se_from_max}{Combination of features that has a mean
cost function that is \code{1.96*SE} from the maximum/minimum mean
cost function.}
}
\description{
Returns the \code{maximum}, \code{se1}, and \code{se2} features from
a completed \code{feature_select} object.
}
\examples{
data  <- wranglr::simdata
feats <- attr(data, "sig_feats")$class
fs <- feature_selection(data, candidate_features = feats,
                        search_type = search_type_forward_model(),
                        model_type  = model_type_lr("class_response"),
                        runs = 2L, folds = 2L)
fs_obj <- Search(fs)
get_fs_features(fs_obj)
}
