% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-markers.R
\name{get_markers}
\alias{get_markers}
\title{Get Feature Selection Markers}
\usage{
get_markers(x)
}
\arguments{
\item{x}{A \code{feature_select} search object.}
}
\value{
A list containing:
\item{max_markers}{Combination of features that gives maximum/minimum
mean cost function.}
\item{markers_1se_from_max}{Combination of features that has a mean
cost function that is one standard error (SE) from the
maximum/minimum mean cost function.}
\item{markers_2se_from_max}{Combination of features that has a mean
cost function that is 1.96*SE from the maximum/minimum mean
cost function.}
}
\description{
Function to return the maximum, se1, and se2 markers from either
forward or backward selection from a \code{feature_select} class object.
}
\examples{
data  <- splyr::sim_adat
feats <- attributes(data)$sig_feats$class
fs <- feature_selection(data, candidate_markers = feats,
                        search_type = searchType_forwardModel(),
                        model_type = model_type_glm("class_response"),
                        stratified = TRUE,
                        cost = "AUC", runs = 2, folds = 2)
fs_obj <- Search(fs)
get_markers(fs_obj)
}
\author{
Stu Field
}
