% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-cross.R
\name{setup_cross}
\alias{setup_cross}
\alias{setup_cross_strat}
\alias{plot_cross}
\title{Set up Cross-Validation Folds}
\usage{
setup_cross(x)

setup_cross_strat(x)

plot_cross(x)
}
\arguments{
\item{x}{A \code{feature_select} class object.}
}
\description{
This S3 method sets up the internal the \emph{non-stratified} cross-validation.
Since these are mostly internal methods, documentation is minimal.
}
\author{
Stu Field
}
